refactor: renames logger filename
