+++ src/features/completion-provider.ts
@@ -4,7 +4,7 @@ import * as vscode from 'vscode';
 import { isPositionInString, FORTRAN_KEYWORDS } from '../lib/helper';
 import { getDeclaredFunctions } from '../lib/functions';
 import { EXTENSION_ID } from '../lib/tools';
-import { Logger } from '../services/logging-service';
+import { Logger } from '../services/logging';
 import intrinsics from './intrinsics.json';
 
 class CaseCoverter {
