+++ src/features/formatting-provider.ts
@@ -6,7 +6,7 @@ import * as which from 'which';
 import * as vscode from 'vscode';
 import * as cp from 'child_process';
 
-import { Logger } from '../services/logging-service';
+import { Logger } from '../services/logging';
 import {
   FORMATTERS,
   EXTENSION_ID,
